UserStatus를 도메인 밸류로 만들고 리포지터리 삭제
