+++ src/main/java/com/sprint/mission/discodeit/domain/user/OnlineStatus.java
@@ -1,0 +1,5 @@
+package com.sprint.mission.discodeit.domain.user;
+
+public enum OnlineStatus {
+    ONLINE, OFFLINE
+}
