--- src/main/java/com/sprint/mission/discodeit/domain/userstatus/OnlineStatus.java
@@ -1,5 +1,0 @@
-package com.sprint.mission.discodeit.domain.userstatus;
-
-public enum OnlineStatus {
-    ONLINE, OFFLINE
-}
