--- src/main/java/com/sprint/mission/discodeit/domain/userstatus/UserStatusRepository.java
@@ -1,20 +1,0 @@
-package com.sprint.mission.discodeit.domain.userstatus;
-
-
-
-import java.util.List;
-import java.util.Optional;
-import java.util.UUID;
-
-public interface UserStatusRepository {
-    void save(UserStatus userStatus);
-
-    void remove(UserStatus userStatus);
-
-    Optional<UserStatus> findById(UUID id);
-
-    Optional<UserStatus> findByUserId(UUID userId);
-
-    List<UserStatus> findAll();
-
-}
