--- src/main/java/com/sprint/mission/discodeit/infrastructure/file/FileUserStatusRepository.java
@@ -1,79 +1,0 @@
-package com.sprint.mission.discodeit.infrastructure.file;
-
-import com.sprint.mission.discodeit.domain.userstatus.UserStatusRepository;
-import com.sprint.mission.discodeit.domain.userstatus.UserStatus;
-
-import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
-import org.springframework.stereotype.Repository;
-
-import java.io.ObjectInputStream;
-import java.io.ObjectOutputStream;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.util.*;
-
-@Repository
-@ConditionalOnProperty(prefix = "discodeit.repository", name = "type", havingValue = "file")
-public class FileUserStatusRepository implements UserStatusRepository {
-
-    private final String FILE_PATH = "data/userstatuses.ser"; // 저장 파일 경로
-
-    private Map<UUID, UserStatus> load() {
-        Path filePath = Path.of(FILE_PATH);
-        if (Files.notExists(filePath)) {
-            return new HashMap<>();
-        }
-        try (ObjectInputStream ois = new ObjectInputStream(Files.newInputStream(filePath))) {
-            return (Map<UUID, UserStatus>) ois.readObject();
-
-        } catch (Exception e) {
-            e.printStackTrace();
-            return new HashMap<>();
-        }
-    }
-
-    private void saveToFile(Map<UUID, UserStatus> data) {
-        Path filePath = Path.of(FILE_PATH);
-        try {
-            try (ObjectOutputStream oos = new ObjectOutputStream(Files.newOutputStream(filePath))) {
-                oos.writeObject(data);
-            }
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-    }
-
-    @Override
-    public void save(UserStatus userStatus) {
-        Map<UUID, UserStatus> store = load();
-        UUID key = userStatus.getId();
-        store.put(key, userStatus);
-        saveToFile(store);
-    }
-
-    @Override
-    public void remove(UserStatus userStatus) {
-        Map<UUID, UserStatus> store = load();
-        UUID key = userStatus.getId();
-        store.remove(key);
-        saveToFile(store);
-    }
-
-    @Override
-    public Optional<UserStatus> findById(UUID id) {
-        Map<UUID, UserStatus> store = load();
-        return Optional.ofNullable(store.get(id));
-    }
-
-    @Override
-    public Optional<UserStatus> findByUserId(UUID userId) {
-        Map<UUID, UserStatus> store = load();
-        return store.values().stream().filter(us->us.getUserId().equals(userId)).findAny();
-    }
-
-    @Override
-    public List<UserStatus> findAll() {
-        Map<UUID, UserStatus> store = load();
-        return List.copyOf(store.values());
-    }
-}
